Move getAgeWarnings into lib
