+++ packages/frontend/amp/components/topMeta/TopMetaNews.tsx
@@ -11,7 +11,7 @@ import { TopMetaExtras } from '@frontend/amp/components/topMeta/TopMetaExtras';
 import { Standfirst } from '@frontend/amp/components/topMeta/Standfirst';
 import { SeriesLink } from '@frontend/amp/components/topMeta/SeriesLink';
 import { getSharingUrls } from '@frontend/lib/sharing-urls';
-import { getAgeWarning } from '@frontend/model/age-warning';
+import { getAgeWarning } from '@frontend/lib/age-warning';
 import { Branding } from '@frontend/amp/components/topMeta/Branding';
 import { StarRating } from '../StarRating';
