+++ packages/frontend/amp/components/topMeta/TopMetaPaidContent.tsx
@@ -11,7 +11,7 @@ import { PaidForBand } from '@frontend/amp/components/topMeta/PaidForBand';
 import { palette } from '@guardian/src-foundations';
 import { textSans, body } from '@guardian/pasteup/typography';
 import { getSharingUrls } from '@frontend/lib/sharing-urls';
-import { getAgeWarning } from '@frontend/model/age-warning';
+import { getAgeWarning } from '@frontend/lib/age-warning';
 
 const headerStyle = css`
     ${textSans(9)};
